modified:   diffdrive_bot/bringup/config/mira_controller.yaml 	modified:   diffdrive_bot/description/urdf/lidar.xacro 	modified:   diffdrive_bot/hardware/include/diffdrive_bot/motor_handler.hpp 	modified:   diffdrive_bot/hardware/src/motor_handler.cpp

+++ diffdrive_bot/hardware/include/diffdrive_bot/motor_handler.hpp
@@ -21,7 +21,7 @@ public:
 
     void initialize(std::string serial_port, int baud, char parity, int data_bit, int stop_bit, int slave_id, int gear_ratio);
     void disconnect();
-    void setSpeed(int value);
+    void setSpeed(double value);
     double getSpeed();
     void brakeEngage();
     void brakeRelease();
